reorganize directories and change bg colors
