+++ packages/dev/sharedUiComponents/.storybook/preview.js
@@ -6,4 +6,21 @@ export const parameters = {
       date: /Date$/,
     },
   },
+  backgrounds: {
+    default: 'normal',
+    values: [
+      {
+        name: 'normal',
+        value: '#e0e0e0'
+      },
+      {
+        name: 'light',
+        value: '#ffffff'
+      },
+      {
+        name: 'dark',
+        value: '#333333'
+      }
+    ]
+  }
 }
