+++ packages/dev/sharedUiComponents/src/components/Icon.scss
